Some improvements regarding compatibility, realism and intelligence
Compatibility with all dimensions of screens: The world is always the
same size and it is only scaled when shown in the screen.
Some parameters have been changed to increase the realism.
Intelligence improved: The players try not to score in their own goal
and have a better attack strategy, best detection of fallen bodies,